Add describe subcommand
